+++ APExtensions/Classes/Core/_Protocols/Occupiable/Occupiable.swift
@@ -10,7 +10,7 @@ import Foundation
 
 // ******************************* MARK: - Declarations
 
-// Take from https://github.com/RxSwiftCommunity/RxOptional/blob/master/Sources/RxOptional/Occupiable.swift
+// Taken from https://github.com/RxSwiftCommunity/RxOptional/blob/master/Sources/RxOptional/Occupiable.swift
 // Originally from here: https://github.com/artsy/eidolon/blob/f95c0a5bf1e90358320529529d6bf431ada04c3f/Kiosk/App/SwiftExtensions.swift#L23-L40
 // Credit to Artsy and @ashfurrow
 // Anything that can hold a value (strings, arrays, etc.)