+++ CarthageSupport/APExtensions/APExtensions.h
@@ -15,6 +15,3 @@ FOUNDATION_EXPORT double APExtensionsVersionNumber;
 FOUNDATION_EXPORT const unsigned char APExtensionsVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <APExtensions/PublicHeader.h>
-
-#import "APExtensionsLoader.h"
-#import "APUtils.h"